[update]MyPlaylistをコンポーネント化
